Boss Blues Driver tone matching with Line 6 Helix.
https://youtu.be/51MNfmw_RT4
https://line6.com/customtone/tone/3678541/
Here is my Boss BD2 patch created by doing a seriously deep dive into how a BD2 functions, and then recreating it in my helix. It has a Valve driver and a Parametric eq. used in series. Both are necessary to mimic the BD2. If you have to level match use the last gain block on the eq as you would a level control on a regular pedal. For info and a Video on how I made this patch please see the youtube channel that bears my name.